+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/IMManager.java
@@ -44,4 +44,13 @@ public class IMManager extends BaseManager {
         );
     }
 
+    public void uploadImage(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.uploadImage(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "path"),
+                value(methodCall, "token"),
+                value(methodCall, "obj")
+        );
+    }
 }