+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/OrganizationManager.java
@@ -1,0 +1,54 @@
+package io.openim.flutter_openim_sdk.manager;
+
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
+import io.openim.flutter_openim_sdk.listener.OnBaseListener;
+import io.openim.flutter_openim_sdk.listener.OnOrganizationListener;
+import open_im_sdk.Open_im_sdk;
+
+public class OrganizationManager extends BaseManager {
+
+    public void setOrganizationListener(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.setOrganizationListener(new OnOrganizationListener());
+    }
+
+    public void getSubDepartment(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getSubDepartment(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "departmentID"),
+                int2long(methodCall, "offset"),
+                int2long(methodCall, "count")
+        );
+    }
+
+    public void getDepartmentMember(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getDepartmentMember(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "departmentID"),
+                int2long(methodCall, "offset"),
+                int2long(methodCall, "count")
+        );
+    }
+
+    public void getUserInDepartment(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getUserInDepartment(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "userID")
+        );
+    }
+
+    public void getDepartmentMemberAndSubDepartment(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getDepartmentMemberAndSubDepartment(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                value(methodCall, "departmentID"),
+                int2long(methodCall, "departmentOffset"),
+                int2long(methodCall, "departmentCount"),
+                int2long(methodCall, "memberOffset"),
+                int2long(methodCall, "memberCount")
+        );
+    }
+}