+++ ios/Classes/SwiftFlutterOpenimSdkPlugin.swift
@@ -10,6 +10,7 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
     let userManger: UserManager
     let signalingManager: SignalingManager
     let workMomentsManager: WorkMomentsManager
+    let organizationManager: OrganizationManager
     
     init(channel: FlutterMethodChannel) {
         self.imManager = IMMananger(channel: channel)
@@ -20,6 +21,7 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
         self.userManger = UserManager(channel: channel)
         self.signalingManager = SignalingManager(channel: channel)
         self.workMomentsManager = WorkMomentsManager(channel: channel)
+        self.organizationManager = OrganizationManager(channel: channel)
     }
     
     public static func register(with registrar: FlutterPluginRegistrar) {
@@ -47,6 +49,8 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
             signalingManager.handleMethod(call: call, result: result)
         case "workMomentsManager":
             workMomentsManager.handleMethod(call: call, result: result)
+        case "organizationManager":
+            organizationManager.handleMethod(call: call, result: result)
         default:
             print("Handle ManagerName Error: \(managerName) not found")
         }