+++ lib/src/listener/organization_listener.dart
@@ -1,0 +1,9 @@
+class OnOrganizationListener {
+  Function()? onOrganizationUpdated;
+
+  OnOrganizationListener({this.onOrganizationUpdated});
+
+  void organizationUpdated() {
+    onOrganizationUpdated?.call();
+  }
+}